+++ Ice Scate/Assets/Materials/BackGround.mat
@@ -42,7 +42,7 @@ Material:
     - _MainTex:
         m_Texture: {fileID: 2800000, guid: fbb7bfb697508bf45952b0b848eef77a, type: 3}
         m_Scale: {x: 1, y: 1}
-        m_Offset: {x: 0, y: 0.07863259}
+        m_Offset: {x: 0, y: 0.8013184}
     - _MetallicGlossMap:
         m_Texture: {fileID: 0}
         m_Scale: {x: 1, y: 1}
