+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/AndroidManifest.xml.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 92c7cf3c4bf87884283197e07710bf95
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
