+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/R.txt
@@ -1,0 +1,8 @@
+int attr adSize 0x1
+int attr adSizes 0x1
+int attr adUnitId 0x1
+int style Theme_IAPTheme 0x1
+int[] styleable AdsAttrs { 0x1, 0x1, 0x1 }
+int styleable AdsAttrs_adSize 0
+int styleable AdsAttrs_adSizes 1
+int styleable AdsAttrs_adUnitId 2
