+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/R.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fef11bcb0e2fcec40b1de615b5bbe9ed
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
