+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/libs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d562403946559d24ba6ed92fd38f21d2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
