+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/libs/classes.jar.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dc99212903ff4b14eb2f67a6838c8929
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
