+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/proguard.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 970243a83b4fd2c419ce56dfb1188f34
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
