+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/project.properties
@@ -1,0 +1,3 @@
+# Project target.
+target=android-9
+android.library=true
