+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/project.properties.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2567a17a9e17835428d926df3f30288c
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
