+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/res.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 945b7f466fef22246be61ec1456533b2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
