+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/res/values.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f91e6bcc6ff74634197ee3cca6186b72
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
