+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/res/values/values.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding='utf-8' standalone='no'?>
+<resources>
+<!-- java/com/google/android/gmscore/integ/client/admob/res/values/iap_style.xml -->
+<eat-comment/>
+<style name="Theme.IAPTheme" parent="android:Theme">
+<item name="android:windowIsTranslucent">true</item>
+<item name="android:windowBackground">@android:color/transparent</item>
+<item name="android:windowContentOverlay">@null</item>
+<item name="android:windowNoTitle">true</item>
+<item name="android:windowIsFloating">true</item>
+<item name="android:backgroundDimEnabled">false</item>
+</style>
+<!-- java/com/google/android/gmscore/integ/client/admob/res/values/ads_attrs.xml -->
+<eat-comment/>
+<declare-styleable name="AdsAttrs">
+<attr name="adSize" format="string"/>
+
+<attr name="adSizes" format="string"/>
+
+<attr name="adUnitId" format="string"/>
+</declare-styleable>
+</resources>
