+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/res/values/values.xml.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 57eb1d1afd216f24db68bc0b2d083572
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
