+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/third_party_licenses.json
@@ -1,0 +1,1 @@
+{"Android Activity Support Library": {"length": 10695, "start": 432}, "AndroidX asynchronous layout inflater library": {"length": 10695, "start": 432}, "AndroidX concurrent futures library": {"length": 10695, "start": 432}, "AndroidX coordinator layout library": {"length": 10695, "start": 432}, "AndroidX cursor adapter library": {"length": 10695, "start": 432}, "AndroidX document file library": {"length": 10695, "start": 432}, "AndroidX drawer layout library": {"length": 10695, "start": 432}, "AndroidX interpolator library": {"length": 10693, "start": 11203}, "AndroidX legacy core UI libraries": {"length": 10695, "start": 432}, "AndroidX local broadcast manager library": {"length": 10693, "start": 11203}, "AndroidX print library": {"length": 10695, "start": 432}, "AndroidX sliding pane layout library": {"length": 10695, "start": 432}, "AndroidX swipe refresh layout library": {"length": 10695, "start": 432}, "AndroidX versioned parcelable library": {"length": 10695, "start": 21938}, "AndroidX view pager library": {"length": 10695, "start": 432}, "Animal Sniffer": {"length": 1096, "start": 32652}, "Checker Framework Annotations": {"length": 1846, "start": 33782}, "Error Prone": {"length": 11357, "start": 35644}, "FlatBuffers": {"length": 11342, "start": 47017}, "Guava JDK5": {"length": 11358, "start": 58403}, "Guava JDK7": {"length": 11362, "start": 69776}, "J2ObjC": {"length": 11358, "start": 58403}, "JSR 250": {"length": 11358, "start": 58403}, "JSR 305": {"length": 1602, "start": 81150}, "safeparcel": {"length": 11358, "start": 58403}, "savedstate": {"length": 10695, "start": 432}}
