+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/third_party_licenses.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fc56817cc282cf54fb13ab4920f2d036
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
