+++ Ice Scate/Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0/third_party_licenses.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4f09fe35d98a59540bc1a7bce93f8d6b
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
