+++ Ice Scate/Assets/Scenes/MainGame.unity
@@ -897,7 +897,18 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 275955817}
   m_OnClick:
     m_PersistentCalls:
-      m_Calls: []
+      m_Calls:
+      - m_Target: {fileID: 1419227760}
+        m_MethodName: OnShowView
+        m_Mode: 1
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
     m_TypeName: UnityEngine.UI.Button+ButtonClickedEvent, UnityEngine.UI, Version=1.0.0.0,
       Culture=neutral, PublicKeyToken=null
 --- !u!114 &275955817
@@ -2135,6 +2146,50 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1035884514}
   m_CullTransparentMesh: 0
+--- !u!1 &1062882311
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1062882312}
+  - component: {fileID: 1062882313}
+  m_Layer: 0
+  m_Name: RewardObject
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &1062882312
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1062882311}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 13
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1062882313
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1062882311}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cebfb7ab7ccaca94bb19707936e9b4a8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  image_game_over: {fileID: 1610354333}
 --- !u!1 &1090765161
 GameObject:
   m_ObjectHideFlags: 0
@@ -2508,6 +2563,7 @@ MonoBehaviour:
   ui_objects_:
   - {fileID: 2057319596}
   - {fileID: 969507332}
+  - {fileID: 1610354333}
   sprites_pause_:
   - {fileID: 21300000, guid: 6f3a0846b246ddf44907eb1bc99c3d34, type: 3}
   - {fileID: 21300000, guid: 1607df49adb0b0c46acdf6ae50418b4e, type: 3}
@@ -2540,6 +2596,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d22ff4b2c834a2246bc61110d5a49f06, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  viewer: {fileID: 1062882313}
   state_: 0
 --- !u!1 &1433900595
 GameObject:
@@ -2798,6 +2855,7 @@ GameObject:
   - component: {fileID: 1610354334}
   - component: {fileID: 1610354336}
   - component: {fileID: 1610354335}
+  - component: {fileID: 1610354337}
   m_Layer: 5
   m_Name: GameOverImage
   m_TagString: Untagged
@@ -2866,6 +2924,25 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1610354333}
   m_CullTransparentMesh: 0
+--- !u!114 &1610354337
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1610354333}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f4cae8513290d7844a4fd19d64162766, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Name: 
+  m_EditorClassIdentifier: 
+  viewer: {fileID: 1062882313}
 --- !u!1 &1625703211
 GameObject:
   m_ObjectHideFlags: 0
