+++ Ice Scate/Assets/Scripts/Admob/ContinueView.cs
@@ -1,0 +1,75 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using GoogleMobileAds.Api;
+using System;
+
+public class ContinueView : MonoBehaviour
+{
+    [SerializeField] private GameObject image_game_over;
+
+    private RewardedAd rewardedAd;
+
+    public void Start()
+    {
+        string adUnitId;
+#if UNITY_ANDROID
+        adUnitId = "ca-app-pub-3940256099942544/5224354917";
+#elif UNITY_IPHONE
+            adUnitId = "ca-app-pub-3940256099942544/1712485313";
+#else
+            adUnitId = "unexpected_platform";
+#endif
+
+        rewardedAd = new RewardedAd(adUnitId);
+
+        //リクエスト挿入
+        rewardedAd.OnUserEarnedReward += HandleUserEarnedReward;
+        rewardedAd.OnAdClosed += HandleRewardedAdClosed;
+
+        // Create an empty ad request.
+        AdRequest request = new AdRequest.Builder().Build();
+        // Load the rewarded ad with the request.
+        rewardedAd.LoadAd(request);
+    }
+
+    public void CreateAndLoadRewardedAd()
+    {
+#if UNITY_ANDROID
+        string adUnitId = "ca-app-pub-3940256099942544/5224354917";
+#elif UNITY_IPHONE
+            string adUnitId = "ca-app-pub-3940256099942544/1712485313";
+#else
+            string adUnitId = "unexpected_platform";
+#endif
+
+        rewardedAd = new RewardedAd(adUnitId);
+
+        //リクエスト挿入
+        rewardedAd.OnUserEarnedReward += HandleUserEarnedReward;
+        rewardedAd.OnAdClosed += HandleRewardedAdClosed;
+
+        // Create an empty ad request.
+        AdRequest request = new AdRequest.Builder().Build();
+        // Load the rewarded ad with the request.
+        rewardedAd.LoadAd(request);
+    }
+
+    public void ShowView()
+    {
+        if (rewardedAd.IsLoaded())
+        {
+            rewardedAd.Show();
+        }
+    }
+
+    public void HandleRewardedAdClosed(object sender, EventArgs args)
+    {
+        CreateAndLoadRewardedAd();
+    }
+
+    public void HandleUserEarnedReward(object sender, Reward args)
+    {
+        image_game_over.SetActive(false);
+    }
+}
