+++ Ice Scate/Assets/Scripts/Admob/ContinueView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cebfb7ab7ccaca94bb19707936e9b4a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
