--- Ice Scate/Assets/Scripts/GameStateManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d22ff4b2c834a2246bc61110d5a49f06
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
