+++ Ice Scate/Assets/Scripts/Managers/PopUpManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cdc0b721fbeffca499bd2fa900e29fdd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
