+++ Ice Scate/Assets/Scripts/MoveObstacle.cs
@@ -40,7 +40,7 @@ public class MoveObstacle : MonoBehaviour
 
     void Update()
     {
-        if(GameStateManager.manager_.state_ == GameStateManager.State.ACTIVE)
+        if(GameManager.manager_.state_ == GameManager.State.ACTIVE)
         {
             switch (type_)
             {
