--- Ice Scate/Assets/Scripts/SoundManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff5915a52bd7aa849a9bda5b1f8efdcf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
