+++ Ice Scate/Assets/Scripts/TogglePause.cs
@@ -27,12 +27,12 @@ public class TogglePause : MonoBehaviour
         {
             SoundManager.instance.PlaySE(1);
             renderer_.sprite = sprites_pause_[1];
-            GameStateManager.manager_.state_ = GameStateManager.State.ACTIVE;
+            GameManager.manager_.state_ = GameManager.State.ACTIVE;
         }
         else
         {
             renderer_.sprite = sprites_pause_[0];
-            GameStateManager.manager_.state_ = GameStateManager.State.PAUSE;
+            GameManager.manager_.state_ = GameManager.State.PAUSE;
         }
     }
 }
