+++ Ice Scate/ProjectSettings/AndroidResolverDependencies.xml
@@ -34,7 +34,7 @@
     <file>Assets/Plugins/Android/com.google.android.gms.play-services-ads-18.3.0.aar</file>
     <file>Assets/Plugins/Android/com.google.android.gms.play-services-ads-base-18.3.0.aar</file>
     <file>Assets/Plugins/Android/com.google.android.gms.play-services-ads-identifier-17.0.0.aar</file>
-    <file>Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0.aar</file>
+    <file>Assets/Plugins/Android/com.google.android.gms.play-services-ads-lite-18.3.0</file>
     <file>Assets/Plugins/Android/com.google.android.gms.play-services-basement-17.0.0.aar</file>
     <file>Assets/Plugins/Android/com.google.android.gms.play-services-gass-18.3.0.aar</file>
     <file>Assets/Plugins/Android/com.google.android.gms.play-services-measurement-base-17.1.0.aar</file>
@@ -45,7 +45,7 @@
     <setting name="androidAbis" value="armeabi-v7a,x86" />
     <setting name="bundleId" value="com.DefaultCompany.IceScate" />
     <setting name="explodeAars" value="True" />
-    <setting name="gradleBuildEnabled" value="True" />
+    <setting name="gradleBuildEnabled" value="False" />
     <setting name="gradleTemplateEnabled" value="False" />
     <setting name="installAndroidPackages" value="True" />
     <setting name="packageDir" value="Assets/Plugins/Android" />
